Fiddling
Seemed like hosting my own annotations.xml was causing the CSE to not work as well? Google is opaque. Fiddling with weighting. My current test case is searching for `faweb` or `grades and comments` -- both of which should have hits on helpdesk.stmarksschool.org -- and weren't seeming to come up in the CSE. Fiddling with Google Sitemaps as well, to see if it's an indexing problem.